[Move] 파일 이동 component/Cummunity/Comment.tsx -> compoentn/Comment/Comment.tsx
